white-tile replacer
I added logic to replace white-ish tiles in a GT450 SVS with a dummy tile (currently a red "x"). The dummy tile is "shared" so that the dummy tile bytes are only represented once in each TIFF directory. For SVS files with a lot of white space (e.g., space between levels of a biopsy), this results in a massive reduction in the size of the SVS file.